requirements.txt
absl-py==2.1.0
backcall==0.2.0
beautifulsoup4==4.12.3
boto3==1.33.13
botocore==1.33.13
cachetools==5.5.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
cycler==0.10.0
Cython==3.0.11
cython-bbox==0.1.5
decorator==5.1.1
easydict==1.13
et-xmlfile==1.1.0
faiss-gpu==1.7.2
filelock==3.12.2
filterpy==1.4.5
fire==0.7.0
flatbuffers==24.3.25
fonttools==4.38.0
fsspec==2023.1.0
gdown==4.7.3
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-auth-oauthlib==0.4.6
grpcio==1.62.3
h5py==3.8.0
huggingface-hub==0.16.4
idna==2.10
imageio==2.31.2
importlib-metadata==6.7.0
ipython==7.34.0
jedi==0.19.1
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
lap==0.4.0
loguru==0.7.2
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
motmetrics==1.4.0
networkx==2.6.3
ninja==1.11.1.1
numpy==1.21.6
oauthlib==3.2.2
onnx==1.8.1
onnx-simplifier==0.3.5
onnxoptimizer==0.3.13
onnxruntime==1.8.0
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
openpyxl==3.0.10
packaging==24.0
pandas==1.3.5
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
prettytable==3.7.0
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybboxes==0.1.6
pycocotools @ git+https://github.com/cocodataset/cocoapi.git@8c9bcc3cf640524c4c20a9c40e89cb6a2f2fa0e9#subdirectory=PythonAPI
Pygments==2.17.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.2
PyWavelets==1.3.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
roboflow==1.1.6
rsa==4.9
s3transfer==0.8.2
sahi==0.11.18
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
shapely==2.0.6
six==1.16.0
smmap==5.0.1
soupsieve==2.4.1
supervision==0.11.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.66.5
traitlets==5.9.0
typing_extensions==4.7.1
ultralytics==8.0.145
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==2.2.3
xmltodict==0.14.2
yacs==0.1.8
yolov5==7.0.13
yolox==0.1.0
zipp==3.15.0
